Implement leaderboard service

+++ leaderboard/entity.go
@@ -3,6 +3,6 @@ package leaderboard
 // User is a struct of leaderboard member detail
 type User struct {
 	Name  string
-	Score int64
-	Rank  int
+	Score float64
+	Rank  int64
 }
